pyproject.toml
[tool.poetry]
name = "test-task"
version = "0.1.0"
description = ""
authors = ["Vlad <vladislav.sytyuk@gmail.com>"]
readme = "README.md"
packages = [{include = "test_task"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.5"
mysqlclient = "^2.2.0"
black = "^23.9.1"
graphene-django = "^3.1.5"
django-filter = "^23.2"
django-graphql-jwt = "^0.4.0"
graphene-file-upload = "^1.3.0"
python-dotenv = "^1.0.0"
openai = "^0.28.0"
langchain = "^0.0.291"
pypdfium2 = "^4.20.0"
chromadb = "^0.4.10"
tiktoken = "^0.5.1"
celery = "^5.3.4"
redis = "^5.0.0"
unstructured = "^0.10.16"
markdown = "^3.4.4"
cryptography = "^41.0.4"
django-celery-beat = "^2.5.0"
django-graphql-ratelimit = "^1.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
